ml_backend/pyproject.toml
[project]
name = "ml-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "pillow>=12.1.0",
    "torch>=2.9.1",
    "transformers>=4.57.6",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
    "ty>=0.0.12",
]
